fix signal hanlder: don't unregister if origin signal handler is SIG_IGN (#1476)
* fix signal hanlder

* fix callable

* fix signal SIGCHLD

* don't close exporter  when process signal event

* fix ut with pytest raises

* add ut

* fix ut

* fix comment

---------

+++ dlrover/python/training_event/error_handler.py
@@ -19,7 +19,6 @@ import traceback
 
 from dlrover.python.training_event.config import Config, get_default_logger
 from dlrover.python.training_event.emitter import Process
-from dlrover.python.training_event.exporter import close_default_exporter
 
 logger = get_default_logger()
 
@@ -80,13 +79,28 @@ class ErrorHandler:
                 content["stack"] = f"get stack failed: {str(e)}"
 
             self._process.instant("exit_sig", content)
-            close_default_exporter()
 
         except Exception as e:
             logger.error(f"process signal {signum} error: {str(e)}")
 
         finally:
-            self.unregister()
+            self._call_original_handler(signum, frame)
+
+    def _call_original_handler(self, signum, frame):
+        """call original handler with signal"""
+
+        # if the handler is callable, call it
+        if callable(self._original_handlers[signum]):
+            self._original_handlers[signum](signum, frame)
+        # if the handler is SIG_IGN or signal is SIGCHLD, do nothing
+        elif (
+            self._original_handlers[signum] == signal.SIG_IGN
+            or signum == signal.SIGCHLD
+        ):
+            return
+        else:
+            if self._registered:
+                self.unregister()
             # call original handler with signal
             os.kill(os.getpid(), signum)
